% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestrat.R
\name{bayestrat}
\alias{bayestrat}
\title{bayestrat}
\usage{
bayestrat(
  data.test,
  data.null = NULL,
  data.pc = NULL,
  race = F,
  plink = F,
  data.null.name = NULL,
  n.pc,
  n.cov,
  priorType = "Laplace",
  priorError = c(2, 1),
  priorLaplace = list(type = "random", G = c(1.4, 0.4), startlambda = NULL),
  priorNormal = list(type = "random", IG = c(1, 1), startvarPC = NULL),
  nIter = 10000,
  burnIn = 1000,
  minAbsBeta = 1e-09,
  alpha = c(0.95, 0.999, 0.9999),
  n.core = 1,
  save.pc = T,
  checkConvergence = F,
  plot.interval = NULL,
  n.chains = 1,
  check.whichPC = c(1, 2, 3),
  check.whichSNP = 1,
  seed = NULL
)
}
\arguments{
\item{data.test}{A data frame consisting of a continuous phenotype, (non-genetic) covariates and genotypes for candidate SNPs as columns, individuals as rows. The column name for race information has to be \code{Race}, if applicable.}

\item{data.null}{(optional) A data frame or matrix of genotypes for null SNPs. Rows represent individuals corresponding to \code{data.test}. Necessary if \code{plink=FALSE}.}

\item{data.pc}{(optional) A data frame or matrix of PC scores. Rows represent individuals corresponding to \code{data.test}. Columns represent scores for a particular PC. PCs to be included in the regression model have to be contained in this dataset. Default is NULL, in which case Bayestrat will conduct PC computation.}

\item{race}{TRUE or FALSE, whether to include race information as a covariate.}

\item{plink}{TRUE or FALSE, whether to call PLINK1.9 (installed under the current working directory) for PC calculation, suitable for large null data. If FALSE, PCs will be calculated within R.}

\item{data.null.name}{(optional) A character of the common filename prefix of the PED and MAP files (.ped, .fam) for the null data. Necessary if \code{plink=TRUE}.}

\item{n.pc}{The total number of PCs to be included into model. Set to 0 if no PC to be added.}

\item{n.cov}{The number of covariates (other than PCs) added to model. Set to 0 if there is no covariates.}

\item{priorType}{"Laplace" or "Normal", whether to assign PC effects with Lapalce or normal priors.}

\item{priorError}{A vector of values of \eqn{a_e} and \eqn{b_e} for the Inverse-Gamma distributed random error term, i.e., \eqn{\sigma_e^2} follows \eqn{IG(a_e,b_e)}.}

\item{priorLaplace}{A list for \code{priorType="Laplace"}, consisitng of:
\describe{
\item{type}{"fixed" or "random", whether to assign \eqn{\lambda} with a fixed constant or not. The prior variance of PC effects is 2/\eqn{\lambda^2}.}
\item{G}{If \code{type="random"}, values of the shape parameter \eqn{a_\lambda} and rate parameter \eqn{b_\lambda} for the Gamma-distributed \eqn{\lambda^2}, i.e. \eqn{\lambda^2} follows \eqn{G(a_\lambda,b_\lambda)}.}
\item{startlambda}{The start value of \eqn{\lambda} in the MCMC iterations. If \code{type="fixed"}, this is the constant value assigned to \eqn{\lambda}.}
}}

\item{priorNormal}{A list for \code{priorType="Normal"}, consisitng of:
\describe{
\item{type}{"fixed" or "random", whether to assign \eqn{\sigma^2_\gamma} (prior variance of PC effects) with a fixed constant or not.}
\item{IG}{If \code{type="random"}, values of \eqn{a_\gamma} and \eqn{b_\gamma} for the Inverse-Gamma distributed \eqn{\sigma^2_\gamma}, i.e. \eqn{\sigma^2_\gamma} follows \eqn{IG(a_\gamma,b_\gamma)}.}
\item{startvarPC}{The start value of \eqn{\sigma^2_\gamma} in the MCMC iterations. If \code{type="fixed"}, this is the constant value assigned to \eqn{\sigma^2_\gamma}.}
}}

\item{nIter}{The number of total iterations. Default is 10000.}

\item{burnIn}{The number of burn-in. Default is 1000.}

\item{minAbsBeta}{If \code{priorType="Laplace"}, this is the minimum absolute value of the PC effects to avoid numeric problems. Default is 1e-9.}

\item{alpha}{A vector of length three consisting of confidence levels. Credible intervals for each level will be calculated.}

\item{n.core}{The number of cores to be used in testing SNPs. Parallelization computing is realized by setting n.core greater than 1. Default is 1.}

\item{save.pc}{TRUE or FALSE, whether to save the calculated PC data to file.}

\item{checkConvergence}{TRUE or FALSE, whether to conduct convergence diagnostics.}

\item{plot.interval}{The interval between two posterior samples that are extracted to draw trace plots.}

\item{n.chains}{The number of chains to run in convergence diagnostics. Default if 1.}

\item{check.whichPC}{On which PCs the convergence diagnostics will be performed. Set to NULL if no diagnostics for PCs.}

\item{check.whichSNP}{On which SNP the convergence diagnostics will be performed.}

\item{seed}{Seed to be used for the implemented MCMC. Default is a random seed.}
}
\value{
A list with the following components:
\describe{
\item{n}{A vector of sample sizes for testing each SNP.}
\item{mu}{The estimated value (posterior mean) of \eqn{\mu}.}
\item{varE}{The estimated value (posterior mean) of \eqn{\sigma_e^2}.}
\item{CIsnp}{A list of three data frames of credible intervals for SNP effects at each confidence level.}
\item{SDsnp}{The estimated standard errors for SNP effects.}
\item{CIcov}{A list of three data frames of credible intervals for covariate effects at each confidence level.}
\item{SDcov}{The estimated standard errors for covariate effects.}
\item{CIpc}{A list of three data frames of credible intervals for PC effects at each confidence level.}
\item{SDpc}{The estimated standard errors for PC effects.}
\item{lambda}{If \code{priorType="Laplace"}, the estimated value of \eqn{\lambda}.}
\item{SDlambda}{If \code{priorType="Laplace"}, the estimated standard error of \eqn{\lambda}.}
\item{varPC}{If \code{priorType="Normal"}, the estimated value of \eqn{\sigma_\gamma^2}.}
\item{SDvarPC}{If \code{priorType="Normal"}, the estimated standard error of \eqn{\sigma_\gamma^2}.}
}
}
\description{
\code{bayestrat} conducts genetic association tests for single nucleotide polymorphisms (SNPs) while accounting for population stratification using principal components (PCs) under the Bayesian framework.
}
\examples{
data(data.test)
data(data.null)
out<-bayestrat(data.test,data.null,race=FALSE,plink=FALSE,n.pc=2,n.cov=2,save.pc=FALSE,seed=30)

}
