% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestratSummary.R
\name{bayestratSummary}
\alias{bayestratSummary}
\title{bayestratSummary}
\usage{
bayestratSummary(output)
}
\arguments{
\item{output}{The output from \code{bayestrat}.}
}
\value{
A list with the following components:
\describe{
\item{snp}{A data frame of the scaled effect (effect size/standard error), effect size, standard error and confidence level for significant SNPs.}
\item{cov}{A similar data frame for significant covariates.}
\item{pc}{A similar data frame for significant PCs.}
}
}
\description{
\code{bayestratSummary} summarizes the output from \code{bayestrat}.
}
