% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcCluster.R
\name{pcCluster}
\alias{pcCluster}
\title{pcCluster}
\usage{
pcCluster(
  data.test,
  data.pc,
  race = T,
  cluster = T,
  cluster.whichPC = c(1, 2, 3),
  n.cluster,
  iter.max = 10,
  algorithm = "Hartigan-Wong",
  plotPC = T,
  plot.whichPC = c(1, 2, 3),
  colorby = "clusters",
  colors = NULL
)
}
\arguments{
\item{data.test}{(opional) A data frame consisting of a continuous phenotype, covariates and genotypes for candidate SNPs as columns, individuals as rows. If \code{race=TRUE}, this is necessary and has to include a column named "Race".}

\item{data.pc}{A data frame or matrix of PC scores. Rows represent individuals. Columns represent PCs.}

\item{race}{TRUE or FALSE, whether to comapre clustering results to race categories.}

\item{cluster}{TRUE or FALSE, whether to cluster individuals by the k-means method.}

\item{cluster.whichPC}{A numerical vector, on which PC(s) the clustering is based. Necessary if \code{cluster=TRUE}.}

\item{n.cluster}{The number of clusters for the k-means method.}

\item{iter.max}{The maximum number of iterations allowed in the k-means algorithm.}

\item{algorithm}{The algorithm used in k-means clustering. One of "Hartigan-Wong", "Lloyd", "Forgy" and "MacQueen". Default is "Hartigan-Wong".}

\item{plotPC}{TRUE or FALSE, whether to generate a 3D PC plot.}

\item{plot.whichPC}{A vector of length three containing which columns of \code{data.pc} to be plotted. Default is (1,2,3).}

\item{colorby}{One of "clusters", "race" or "none", whether to color the 3D PC plot by the clustering result, the race information or no color.}

\item{colors}{A vector of color names to be used for plots.}
}
\value{
If \code{cluster=T}, returns the object from \code{kmeans}, and a confusion matrix is also included if \code{race=T}.
}
\description{
\code{pcCluster} performs individual clustering based on PC scores and generates a 3D plot in a pop-up window.
}
\examples{
data(data.test)
data(pc)
out<-pcCluster(data.test,data.pc=pc,race=TRUE,cluster=TRUE,n.cluster=4,
plotPC=FALSE,plot.whichPC=c(1,2,3),colorby="clusters")


}
